CCP people/community detection/requirements.txt
networkx
matplotlib
python-igraph
node2vec
sklearn
gensim